Required Files
This is a Makeup Landing Page, that I have created as a part of one of the Task given by Mussut Private Limited before the start of the actual Internship, as a part of Selection Process. I am glad, that Mussut Private Limited has given me this opportunity to present my piece of skill to them. To be precise, I was supposed to create a Responsive Footer, but I chose to integrate it with a Landing Page.